chore(js): use global mocks
