+++ js/__mocks__/@react-native-community/audio-toolkit.js
@@ -1,0 +1,17 @@
+export class Player {
+	prepare(cb) {
+		if (cb) {
+			cb()
+		}
+	}
+	play(cb) {
+		if (cb) {
+			cb()
+		}
+	}
+	seek(_, cb) {
+		if (cb) {
+			cb()
+		}
+	}
+}
