+++ js/__mocks__/@shakebugs/react-native-shake.js
@@ -1,0 +1,3 @@
+export default {
+	start: jest.fn(() => {}),
+}
