+++ js/__mocks__/react-native-screens.js
@@ -1,0 +1,14 @@
+const View = require('react-native').View
+
+export const enableScreens = jest.fn()
+export const ScreenContainer = View
+export const Screen = View
+export const NativeScreen = View
+export const NativeScreenContainer = View
+export const ScreenStack = View
+export const ScreenStackHeaderConfig = View
+export const ScreenStackHeaderSubview = View
+export const ScreenStackHeaderRightView = View
+export const ScreenStackHeaderLeftView = View
+export const ScreenStackHeaderTitleView = View
+export const ScreenStackHeaderCenterView = View
