+++ js/__mocks__/react-native-share.js
@@ -1,0 +1,1 @@
+export default jest.fn()
