+++ js/package.json
@@ -106,7 +106,11 @@
     "testPathIgnorePatterns": [
       "/node_modules/",
       "/packages/web-dev-app/",
-      "/android/.gomobile-cache"
-    ]
+      "/android/.gomobile-cache/"
+    ],
+    "moduleNameMapper": {
+      "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/assetsTransformer.js",
+      "\\.(css|less)$": "<rootDir>/assetsTransformer.js"
+    }
   }
 }
